Move to internal/elasticsearch package
